@@ -1,13 +1,12 @@
 [build-system]
-requires = ["poetry-core>=1.2.0", "poetry-dynamic-versioning"]
-build-backend = "poetry_dynamic_versioning.backend"
+requires = ["poetry-core>=2.0.0,<3.0.0", "poetry-dynamic-versioning"]
+build-backend = "poetry.core.masonry.api"
 
-[tool.poetry]
+[project]
 name = "jmapc"
-version = "0.0.0"
 description = "JMAP client library for Python"
 license = "GPL-3.0-or-later"
-authors = ["Stephen Kent <smkent@smkent.net>"]
+authors = [ { name = "Stephen Kent", email = "smkent@smkent.net" } ]
 readme = "README.md"
 repository = "https://github.com/smkent/jmapc"
 classifiers = [
@@ -18,19 +17,29 @@ classifiers = [
     "Topic :: Software Development :: Libraries :: Python Modules",
     "Typing :: Typed",
 ]
+requires-python = ">=3.9,<4.0"
+dynamic = [ "version" ]
+dependencies = [
+    "brotli (>=1.0.9)",
+    "dataclasses-json",
+    "python-dateutil",
+    "requests",
+    "sseclient",
+]
 include = [
     "examples",
 ]
 
-[tool.poetry.dependencies]
-python = "^3.8"
-brotli = ">=1.0.9"
-dataclasses-json = "*"
-python-dateutil = "*"
-requests = "*"
-sseclient = "*"
+[project.scripts]
 
-[tool.poetry.dev-dependencies]
+[tool.poetry]
+requires-poetry = ">=2.0"
+version = "0.0.0"
+
+[tool.poetry.requires-plugins]
+poetry-dynamic-versioning = { version = ">=1.0.0,<2.0.0", extras = ["plugin"] }
+
+[tool.poetry.group.dev.dependencies]
 bandit = {extras = ["toml"], version = "*"}
 black = "*"
 cruft = "*"
