@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from dataclasses import dataclass
-from typing import Any, Dict, List, Optional, Type, cast
+from typing import Any, Optional, cast
 
 from .serializer import Model
 
@@ -9,11 +9,11 @@ __all__ = ["Error", "ServerFail"]
 
 
 class ErrorCollector(Model):
-    error_types: Dict[str, Type[Error]] = {}
+    error_types: dict[str, type[Error]] = {}
 
     @classmethod
     def __init_subclass__(cls) -> None:
-        error_class = cast(Type["Error"], cls)
+        error_class = cast(type["Error"], cls)
         type_attr = getattr(error_class, "_type", None)
         if type_attr:
             ErrorCollector.error_types[type_attr] = error_class
@@ -62,7 +62,7 @@ class CannotCalculateChanges(Error):
 @dataclass
 class InvalidArguments(Error):
     _type = "invalidArguments"
-    arguments: Optional[List[str]] = None
+    arguments: Optional[list[str]] = None
     description: Optional[str] = None
 
 
