@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from dataclasses import dataclass, field
-from typing import Dict, List, Optional
+from typing import Optional
 
 from dataclasses_json import config
 
@@ -32,15 +32,15 @@ class IdentityGet(IdentityBase, Get):
 
 @dataclass
 class IdentityGetResponse(IdentityBase, GetResponse):
-    data: List[Identity] = field(metadata=config(field_name="list"))
+    data: list[Identity] = field(metadata=config(field_name="list"))
 
 
 @dataclass
 class IdentitySet(IdentityBase, Set):
-    create: Optional[Dict[str, Identity]] = None
+    create: Optional[dict[str, Identity]] = None
 
 
 @dataclass
 class IdentitySetResponse(IdentityBase, SetResponse):
-    created: Optional[Dict[str, Optional[Identity]]]
-    updated: Optional[Dict[str, Optional[Identity]]]
+    created: Optional[dict[str, Optional[Identity]]]
+    updated: Optional[dict[str, Optional[Identity]]]
