@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from dataclasses import dataclass
-from typing import List, Optional
+from typing import Optional
 
 from ..serializer import Model
 from .models import EmailAddress
@@ -12,7 +12,7 @@ class Identity(Model):
     name: str
     email: str
     reply_to: Optional[str]
-    bcc: Optional[List[EmailAddress]]
+    bcc: Optional[list[EmailAddress]]
     text_signature: Optional[str]
     html_signature: Optional[str]
     may_delete: bool
